LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-cors-xhr.https-expected.txt

Blocked access to external URL https://www1.localhost:9443/service-workers/service-worker/resources/fetch-access-control-login.html
CONSOLE MESSAGE: Unable to post message to https://www1.localhost:9443. Recipient has origin https://localhost:9443.



Harness Error (TIMEOUT), message = null

TIMEOUT Verify CORS XHR of fetch() in a Service Worker Test timed out
